Moved InputNumberView and createLabeledInputNumber() from ckeditor5-list to ckeditor5-ui.
