+++ packages/ckeditor5-ui/src/index.js
@@ -34,6 +34,7 @@ export { default as FocusCycler } from './focuscycler';
 export { default as IconView } from './icon/iconview';
 export { default as InputView } from './input/inputview';
 export { default as InputTextView } from './inputtext/inputtextview';
+export { default as InputNumberView } from './inputnumber/inputnumberview';
 
 export { default as IframeView } from './iframe/iframeview';
